Work from 2/28/2022
Started In On Genesis 1 verses 1 and 2.
Concepts  explored God as class representable entity that creates the Earth and Heavens. I thought the Earth and heavens would be interestingly represented via a class Called Spacetime that had members representing the laws of the universe. Also Planets and etc could be added as data members after I initiate those classes